msm: kgsl: Wake GPU upon receiving an ioctl rather than upon touch input
Waking the GPU upon touch wastes power when the screen is being touched
in a way that does not induce animation or any actual need for GPU usage.
Instead of preemptively waking the GPU on touch input, wake it up upon
receiving a IOCTL_KGSL_GPU_COMMAND ioctl since it is a sign that the GPU
will soon be needed.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[md: Adapted for A418,we don't have IOCTL_KGSL_GPU_COMMAND
instead use IOCTL_KGSL_CMDSTREAM_FREEMEMONTIMESTAMP_CTXTID
which also show soon or later GPU need to wake out of sleep state]
Signed-off-by: mortezadadgar <mortezadadgar97@gmail.com>